unique names for shots/assets

+++ byuam/project.py
@@ -68,22 +68,38 @@ class Project:
 			return None
 		return Shot(filepath)
 
+	def get_body(self, name):
+		"""
+		returns the body object associated with the given name.
+		name -- the name of the body
+		"""
+		body = self.get_shot(name)
+		if body is None:
+			body = self.get_asset(name)
+		return body
+
+	def _create_body(self, name, bodyobj):
+		name = pipeline_io.alphanumeric(name)
+		filepath = os.path.join(bodyobj.get_parent_dir(), name)
+		if name in self.list_bodies():
+			raise EnvironmentError("body already exists: "+filepath)
+		if not pipeline_io.mkdir(filepath):
+			raise OSError("couldn't create body directory: "+filepath)
+		datadict = bodyobj.create_new_dict(name)
+		pipeline_io.writefile(os.path.join(filepath, bodyobj.PIPELINE_FILENAME), datadict)
+		new_body = bodyobj(filepath)
+		for dept in bodyobj.default_departments():
+			pipeline_io.mkdir(os.path.join(filepath, dept))
+			new_body.create_element(dept, Element.DEFAULT_NAME)
+		return new_body
+
 	def create_asset(self, name):
 		"""
 		creates a new shot with the given name, and returns the resulting shot object.
 		If a shot with that name already exists, raises EnvironmentError.
 		name -- the name of the new shot to create
 		"""
-		filepath = os.path.join(self._env.get_assets_dir(), name)
-		if not pipeline_io.mkdir(filepath):
-			raise EnvironmentError("asset already exists: "+filepath)
-		datadict = Asset.create_new_dict(name)
-		pipeline_io.writefile(os.path.join(filepath, Body.PIPELINE_FILENAME), datadict)
-		new_asset = Asset(filepath)
-		for dept in Department.FRONTEND:
-			pipeline_io.mkdir(os.path.join(filepath, dept))
-			new_asset.create_element(dept, Element.DEFAULT_NAME)
-		return new_asset
+		return self._create_body(name, Asset)
 
 	def create_shot(self, name):
 		"""
@@ -91,18 +107,9 @@ class Project:
 		If a shot with that name already exists, raises EnvironmentError.
 		name -- the name of the new shot to create
 		"""
-		filepath = os.path.join(self._env.get_shots_dir(), name)
-		if not pipeline_io.mkdir(filepath):
-			raise EnvironmentError("shot already exists: "+filepath)
-		datadict = Shot.create_new_dict(name)
-		pipeline_io.writefile(os.path.join(filepath, Body.PIPELINE_FILENAME), datadict)
-		new_shot = Shot(filepath)
-		for dept in Department.BACKEND:
-			pipeline_io.mkdir(os.path.join(filepath, dept))
-			new_shot.create_element(dept, Element.DEFAULT_NAME)
-		return new_shot
+		return self._create_body(name, Shot)
 
-	def _list_bodies(self, filepath):
+	def _list_bodies_in_dir(self, filepath):
 		dirlist = os.listdir(filepath)
 		assetlist = []
 		for assetdir in dirlist:
@@ -116,13 +123,19 @@ class Project:
 		"""
 		returns a list of strings containing the names of all assets in this project
 		"""
-		return self._list_bodies(self._env.get_assets_dir())
+		return self._list_bodies_in_dir(self._env.get_assets_dir())
 
 	def list_shots(self):
 		"""
 		returns a list of strings containing the names of all shots in this project
 		"""
-		return self._list_bodies(self._env.get_shots_dir())
+		return self._list_bodies_in_dir(self._env.get_shots_dir())
+
+	def list_bodies(self):
+		"""
+		returns a list of strings containing the names of all bodies (assets and shots)
+		"""
+		return self.list_assets() + self.list_shots()
 
 	def is_checkout_dir(self, path):
 		"""
@@ -144,10 +157,12 @@ class Project:
 		"""
 		delete the given shot
 		"""
-		shutil.rmtree(os.path.join(self.get_shots_dir(), shot))
+		if shot in self.list_shots():
+			shutil.rmtree(os.path.join(self.get_shots_dir(), shot))
 
 	def delete_asset(self, asset):
 		"""
 		delete the given asset
 		"""
-		shutil.rmtree(os.path.join(self.get_asset_dir(), asset))
+		if asset in self.list_assets():
+			shutil.rmtree(os.path.join(self.get_asset_dir(), asset))
